serverside/marshmallow/requirements.txt
appdirs==1.4.3
Django==1.5
docutils==0.13.1
funcsigs==1.0.2
futures==3.1.1
gunicorn==19.7.1
IndicoIo==1.0.8
mock==1.3.0
mongoengine==0.13.0
oauthlib==2.0.2
olefile==0.44
packaging==16.8
pbr==2.1.0
Pillow==4.1.0
pymongo==3.4.0
pyparsing==2.2.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
statistics==1.0.3.5
tweepy==3.5.0
tweet-preprocessor==0.5.0
whitenoise==3.3.0
